Do not wait for the account_data stream to update when removing account data
In the case of attempting to remove an entry from a user or room's
account data, a request may come in to remove an entry that does not
exist. In this case, we do not update the maximum account_data stream,
as we haven't modified any data. We also do not notify any other
workers, synapse modules, or other listeners that a change occurred.

In the case of using workers with Synapse, the master process will task
a worker (if flagged as an account data write) with carrying out the
account data deletion. When it does so, it initiates a replication
command tied to the account data stream. By default, these commands
expect that the stream will be incremented as a result of this command,
and thus wait for the stream to be updated before considering the task
completed. You can see the discrepancy here.

So we tell the master not to wait on the stream to be updated instead. I
*think* this is safe, as it will still wait for the replication request
to complete, and in the case of an actual account data deletion
occurring, the worker should not respond to the replication request
until the deletion has happened and listeners are notified.